google_calendar/requirements.txt
# Dashy Calendar Feed Requirements
# Install with: pip install -r requirements.txt

# Google Calendar API
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Environment variables
python-dotenv>=0.19.0

# Optional: For better error handling and logging
requests>=2.25.0